add option for bindings to choose what the source is
when binding, youd want the two values to be equal, so youd always want
callable to be called, so ive removed that option. You can instead
choose if the source of the value should come from this observer
property or it should come from the valued stored at the bound object.
Default behaviour is that it should come from this observer property.